requirements.txt
flask
openai
requests
gunicorn
flask-cors
psycopg2-binary
